✊🍹 Checkpoint ./commands/choose.js:2501470/112 ./commands/ask.js:2501470/56 ./commands/ping.js:2501470/56 ./commands/randchar.js:2501470/56 ./commands/say.js:2501470/56 ./commands/setchannels.js:2501470/56 ./events/message.js:2501470/4

+++ commands/say.js
@@ -1,4 +1,5 @@
 exports.run = (client, message, args) => {
+  if(message.guild.id != process.env.HOMESERVER)return;
 	var text = args.join(" ");
     message.channel.send(text).then(message.delete()).catch(console.error);
 }
